oData Implementation + Logic to Handle Reboot/Crash of Nod Device
When you subscribe/unsubscribe to events in the app, this now happens
behind the scenes using oData.

Added logic to keep track of which events each Nod device was
subscribed to in case of reboot or crash.

On reboot, the Nod device will call didDisconnectFromNod method and
then the app will immediately try to connect back to the Nod device.

Added 2 additional tactile buttons in ButtonEvent for Nod Backspin.

Change-Id: Ie27013615e958f10c9866fcddd538115d729565c
Reviewed-on: https://r.nod.com/4606
Tested-by: Jenkins
Reviewed-by: Khwaab Dave <khwaab@nod-labs.com>